// web/src/api/models/TypesLearnedRoute.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesLearnedRoute
 */
export interface TypesLearnedRoute {
    /**
     * 
     * @type {number}
     * @memberof TypesLearnedRoute
     */
    distance?: number;
    /**
     * 
     * @type {string}
     * @memberof TypesLearnedRoute
     */
    interface?: string;
    /**
     * 
     * @type {number}
     * @memberof TypesLearnedRoute
     */
    metric?: number;
    /**
     * 
     * @type {string}
     * @memberof TypesLearnedRoute
     */
    nexthop?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLearnedRoute
     */
    prefix: string;
    /**
     * 
     * @type {string}
     * @memberof TypesLearnedRoute
     */
    protocol: string;
}

/**
 * Check if a given object implements the TypesLearnedRoute interface.
 */
export function instanceOfTypesLearnedRoute(value: object): value is TypesLearnedRoute {
    if (!('prefix' in value) || value['prefix'] === undefined) return false;
    if (!('protocol' in value) || value['protocol'] === undefined) return false;
    return true;
}

export function TypesLearnedRouteFromJSON(json: any): TypesLearnedRoute {
    return TypesLearnedRouteFromJSONTyped(json, false);
}

export function TypesLearnedRouteFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesLearnedRoute {
    if (json == null) {
        return json;
    }
    return {
        
        'distance': json['distance'] == null ? undefined : json['distance'],
        'interface': json['interface'] == null ? undefined : json['interface'],
        'metric': json['metric'] == null ? undefined : json['metric'],
        'nexthop': json['nexthop'] == null ? undefined : json['nexthop'],
        'prefix': json['prefix'],
        'protocol': json['protocol'],
    };
}

export function TypesLearnedRouteToJSON(json: any): TypesLearnedRoute {
    return TypesLearnedRouteToJSONTyped(json, false);
}

export function TypesLearnedRouteToJSONTyped(value?: TypesLearnedRoute | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'distance': value['distance'],
        'interface': value['interface'],
        'metric': value['metric'],
        'nexthop': value['nexthop'],
        'prefix': value['prefix'],
        'protocol': value['protocol'],
    };
}